1.3.4 - template snags

+++ app/src/main/java/com/atakmap/android/soothsayer/models/linksmodel/LinkRequest.kt
@@ -36,7 +36,7 @@ data class LinkRequest(
 
 data class Point(
     val markerId: String,
-    val alt: Int?,
+    val alt: Double?,
     val lat: Double?,
     val lon: Double?
 )
